examples/contrib/detection/requirements.txt
pytorch-ignite
torchvision
tqdm==4.61.2
aim==3.8.0
pycocotools==2.0.4
albumentations==1.1.0
fire==0.4.0
